+++ pyproject.toml
@@ -125,3 +125,16 @@ disallow_untyped_defs = true
 no_implicit_optional = true
 show_error_codes = true
 warn_unused_ignores = true
+
+[tool.pytest.ini_options]
+rootdir = "."  # Ensures pytest always treats project root as the root
+pythonpath = [
+    "tests",
+    "src/icefabric_api",
+    "src/icefabric_tools"
+]
+testpaths = [
+    "src/icefabric_api/tests",
+    "src/icefabric_tools/tests",
+    "tests"
+]
