--- src/icefabric_api/tests/data/usgs_01010000_data_from_20211231_1400_to_20220101_1400.csv
@@ -1,26 +1,0 @@
-q_cms,time
-17.698,2021-12-31 14:00:00
-17.698,2021-12-31 15:00:00
-17.698,2021-12-31 16:00:00
-17.698,2021-12-31 17:00:00
-17.698,2021-12-31 18:00:00
-17.698,2021-12-31 19:00:00
-17.698,2021-12-31 20:00:00
-17.698,2021-12-31 21:00:00
-17.414831,2021-12-31 22:00:00
-17.414831,2021-12-31 23:00:00
-17.414831,2022-01-01 00:00:00
-17.414831,2022-01-01 01:00:00
-17.414831,2022-01-01 02:00:00
-17.414831,2022-01-01 03:00:00
-17.414831,2022-01-01 04:00:00
-17.414831,2022-01-01 05:00:00
-17.414831,2022-01-01 06:00:00
-17.414831,2022-01-01 07:00:00
-17.414831,2022-01-01 08:00:00
-17.414831,2022-01-01 09:00:00
-17.15998,2022-01-01 10:00:00
-17.15998,2022-01-01 11:00:00
-17.15998,2022-01-01 12:00:00
-17.15998,2022-01-01 13:00:00
-17.15998,2022-01-01 14:00:00
