+++ src/icefabric_tools/tests/test_subset.py
@@ -7,7 +7,7 @@ from pyiceberg.catalog import Catalog
 from icefabric_tools import IdType, subset
 
 
-@pytest.mark.integration
+@pytest.mark.slow
 def test_subset_hl_uri(hydrofabric_catalog: Catalog, gauge_hf_uri: str, testing_dir: Path, tmp_path: Path):
     """Tests all subset gauges in the sample data using hl_uri"""
     subset(
