+++ tests/README.md
@@ -1,0 +1,7 @@
+# Pytests
+
+This folder is for all testing data so the global confest can pick it up. This allows all tests in the namespace packages to share the same scope without having to reference one another in tests
+
+To run tests, run `pytest -s` from project root.
+
+To run the subsetter tests, run `pytest --run-slow` as these tests take some time. Otherwise, they will be skipped
